Rename project to daisy-N2O + fix hardcoded win reg key name
We need a workflow where we can test new components/models. This
should probably be a plugin based architecture, but for now we just
use a different project name.

On windows the project name is used as part of the registry key that
stores the location of daisy lib/ and sample/ directories. The key
used in path.C was previously partly hardcoded and thus failed when
the project name is changed.